requirements.txt
streamlit==1.32.2
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.29
chromadb==0.4.24
openai==1.23.2
python-dotenv==1.0.1
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pydantic==2.7.1
typing-extensions==4.11.0
supabase==2.4.5
pandas==2.2.2
fredapi==0.5.0
requests==2.31.0 